requirements.txt
requests>=2.28
python-dotenv>=1.0.0
langchain-ollama>=0.0.1
langchain-core>=0.0.1

# Notes:
# - `langchain-ollama` and `langchain-core` are referenced as imports in the
#   codebase (`langchain_ollama`, `langchain_core`). If those packages have
#   different PyPI names, adjust accordingly.
# - `ollama` is an external CLI/service and is not a Python package.